@@ -1,10 +1,14 @@
 <script lang="ts">
     import { PlainTableCssTheme, type GridTheme, PagingData } from '$lib/index.js';
 
-    export let theme:GridTheme = PlainTableCssTheme;
-    export let paging:PagingData = new PagingData();
+    interface Props {
+        theme?: GridTheme;
+        paging?: PagingData;
+    }
+
+    let { theme = PlainTableCssTheme, paging = $bindable(new PagingData()) }: Props = $props();
 </script>
 
-<svelte:component this={theme.footer} bind:paging >
-    <svelte:component this={theme.paging} bind:paging />
-</svelte:component>
+<theme.footer bind:paging >
+    <theme.paging bind:paging />
+</theme.footer>
