@@ -1,4 +1,6 @@
 <script lang="ts">
+    import { run } from 'svelte/legacy';
+
     import { PlainTableCssTheme } from '$lib/index.js';
 	import { GridFunctions } from "../GridFunctions.js";
     import type { GridColumn, GridFilter, GridTheme, GroupHeader, PagingDataInternal } from "$lib/types/index.js";
@@ -7,30 +9,45 @@
     type T = $$Generic<any>;
     type ExpandedGroups = { [value:string] : boolean };
 
-    export let data: Iterable<T> | ArrayLike<T> = [];
-    export let dataUnpaged: Iterable<T> | ArrayLike<T> = []; // same as data, but nut cut for paging. Used for exports
-    export let columns: GridColumn<T>[] = [];
-    export let filters: GridFilter[] = [];
 
-    export let groupBy = ""; // Default grouping column
-    export let sortByColumn = "";
-    export let sortOrder = 1; // 1 for ascending, -1 for descending
 
-    export let showCheckboxes = false;
-    export let groupsExpandedDefault = true;
-    export let selectedRows: T[] = [];
 
-    export let theme: GridTheme = PlainTableCssTheme;
-    export let paging = new PagingData();
+    interface Props {
+        data?: Iterable<T> | ArrayLike<T>;
+        dataUnpaged?: Iterable<T> | ArrayLike<T>; // same as data, but nut cut for paging. Used for exports
+        columns?: GridColumn<T>[];
+        filters?: GridFilter[];
+        groupBy?: string; // Default grouping column
+        sortByColumn?: string;
+        sortOrder?: number; // 1 for ascending, -1 for descending
+        showCheckboxes?: boolean;
+        groupsExpandedDefault?: boolean;
+        selectedRows?: T[];
+        theme?: GridTheme;
+        paging?: any;
+    }
 
-    let sortOrderSecondary = 1; // 1 for ascending, -1 for descending
-    let expandedGroups: ExpandedGroups = {};
-    let gridData: T[] = [];
-    let groupHeaders: GroupHeader<T>[] = [];
-    let groupHeadersUnpaged: GroupHeader<T>[] = [];
+    let {
+        data = [],
+        dataUnpaged = $bindable([]),
+        columns = $bindable([]),
+        filters = [],
+        groupBy = "",
+        sortByColumn = $bindable(""),
+        sortOrder = $bindable(1),
+        showCheckboxes = false,
+        groupsExpandedDefault = true,
+        selectedRows = $bindable([]),
+        theme = PlainTableCssTheme,
+        paging = $bindable(new PagingData())
+    }: Props = $props();
 
-    $: fulldata = Array.from(data);
-    $: columns, assignAutoColumns();
+    let sortOrderSecondary = $state(1); // 1 for ascending, -1 for descending
+    let expandedGroups: ExpandedGroups = $state({});
+    let gridData: T[] = $state([]);
+    let groupHeaders: GroupHeader<T>[] = $state([]);
+    let groupHeadersUnpaged: GroupHeader<T>[] = $state([]);
+
     function assignAutoColumns() {
         if (columns.length > 0 || fulldata.length == 0) {
             return;
@@ -46,37 +63,14 @@
         });
     }
 
-    let uniqueRowIds: {row: T, id: string}[] = [];
-    $: uniqueRowIds = fulldata.map((row) => {
-        const id = Math.random().toString(36).substr(2, 9);
-        return { row: row, id: id  };
-    }), resetSelectedRows();
+    let uniqueRowIds: {row: T, id: string}[] = $state([]);
     function resetSelectedRows() {
         selectedRows = []; // resetting the selected rows when fulldata changes
     }
     
-    $: columns.forEach((col) => {
-        if (col.visible === undefined) {
-            col.visible = true;
-        }
-    })
 
-    $: activeFilters = filters.filter(x => x.active);
 
-    $: grid = new GridFunctions<T>()
-        .init(fulldata)
-        .applyFilters(filters, columns)
-        .sortBy(sortByColumn, sortOrder, groupBy, sortOrderSecondary, columns)
-        .groupBy(groupBy,expandedGroups, groupsExpandedDefault, columns)
-        .processPaging(groupBy, paging.currentPage, paging.itemsPerPage);
 
-    $: {
-        gridData = grid.data;
-        dataUnpaged = grid.dataUnpaged;
-        groupHeaders = grid.groupHeaders;
-        groupHeadersUnpaged = grid.groupHeadersUnpaged;
-        updatePaging();
-    }
 
     function updatePaging() {
         (paging as PagingDataInternal).totalResults = grid.dataUnpaged.length;
@@ -144,9 +138,40 @@
             selectedRows = [...selectedRows, ...toAdd];
         }
     }
+    let fulldata = $derived(Array.from(data));
+    run(() => {
+        columns, assignAutoColumns();
+    });
+    run(() => {
+        uniqueRowIds = fulldata.map((row) => {
+            const id = Math.random().toString(36).substr(2, 9);
+            return { row: row, id: id  };
+        }), resetSelectedRows();
+    });
+    run(() => {
+        columns.forEach((col) => {
+            if (col.visible === undefined) {
+                col.visible = true;
+            }
+        })
+    });
+    let activeFilters = $derived(filters.filter(x => x.active));
+    let grid = $derived(new GridFunctions<T>()
+        .init(fulldata)
+        .applyFilters(filters, columns)
+        .sortBy(sortByColumn, sortOrder, groupBy, sortOrderSecondary, columns)
+        .groupBy(groupBy,expandedGroups, groupsExpandedDefault, columns)
+        .processPaging(groupBy, paging.currentPage, paging.itemsPerPage));
+    run(() => {
+        gridData = grid.data;
+        dataUnpaged = grid.dataUnpaged;
+        groupHeaders = grid.groupHeaders;
+        groupHeadersUnpaged = grid.groupHeadersUnpaged;
+        updatePaging();
+    });
 </script>
 
-<svelte:component this={theme.grid.container}>
+<theme.grid.container>
     <colgroup>
         {#if showCheckboxes}
             <col span="1">
@@ -157,32 +182,32 @@
             {/if}
         {/each}
     </colgroup>
-    <svelte:component this={theme.grid.header.container}>
-        <svelte:component this={theme.grid.header.row}>
+    <theme.grid.header.container>
+        <theme.grid.header.row>
             {#if showCheckboxes}
-                <svelte:component this={theme.grid.header.checkbox} checked={activeFilters.length == 0 ? fulldata.every(item => selectedRows.includes(item)) : Array.from(dataUnpaged).every(item => selectedRows.includes(item))} onChange={() => {toggleHeaderCheckbox()}} />
+                <theme.grid.header.checkbox checked={activeFilters.length == 0 ? fulldata.every(item => selectedRows.includes(item)) : Array.from(dataUnpaged).every(item => selectedRows.includes(item))} onChange={() => {toggleHeaderCheckbox()}} />
             {/if}
             {#each columns.filter(x => x.visible != false) as col (col.key)}
                 {#if groupBy != col.key}
                     {@const sortable = col.sortable === undefined || col.sortable === true}
-                    <svelte:component this={theme.grid.header.content} title={col.title} sortable={sortable} onClick={() => sortable && handleSort(col.key)}>
+                    <theme.grid.header.content title={col.title} sortable={sortable} onClick={() => sortable && handleSort(col.key)}>
                         {#if sortable}
-                            <svelte:component this={theme.grid.header.sortIndicator} isSorted={sortByColumn == col.key} isDescending={(groupBy && sortOrderSecondary == 1) || (!groupBy && sortOrder == 1)} />
+                            <theme.grid.header.sortIndicator isSorted={sortByColumn == col.key} isDescending={(groupBy && sortOrderSecondary == 1) || (!groupBy && sortOrder == 1)} />
                         {/if}
-                    </svelte:component>
+                    </theme.grid.header.content>
                 {/if}
             {/each}
-        </svelte:component>
-    </svelte:component>
-    <svelte:component this={theme.grid.body.container}>
+        </theme.grid.header.row>
+    </theme.grid.header.container>
+    <theme.grid.body.container>
         {#if groupBy}
             {#each groupHeaders as header, groupIndex (header.groupKey)}
                 {@const unpagedHeader = groupHeadersUnpaged.find(x => x.groupKey == header.groupKey)}
-                <svelte:component this={theme.grid.groupby.container} isSelected={showCheckboxes && unpagedHeader?.data.every(item => selectedRows.includes(item))}>
+                <theme.grid.groupby.container isSelected={showCheckboxes && unpagedHeader?.data.every(item => selectedRows.includes(item))}>
                     {#if showCheckboxes}
-                        <svelte:component this={theme.grid.groupby.checkbox} checked={unpagedHeader?.data.every(item => selectedRows.includes(item))} onChange={() => { toggleGroupCheckbox(header); }} index={groupIndex} />
+                        <theme.grid.groupby.checkbox checked={unpagedHeader?.data.every(item => selectedRows.includes(item))} onChange={() => { toggleGroupCheckbox(header); }} index={groupIndex} />
                     {/if}
-                    <svelte:component this={theme.grid.groupby.cell} colspan={columns.length-1} onToggle={() => toggleGroup(header)} isExpanded={header.expanded}>
+                    <theme.grid.groupby.cell colspan={columns.length-1} onToggle={() => toggleGroup(header)} isExpanded={header.expanded}>
                         {@const col = columns.find(x => x.visible != false && x.key == groupBy)}
                         {#if col}
                             {@const value = header.titleData}
@@ -190,69 +215,78 @@
     
                             {#if typeof value === 'object'}
                                 {#if Object.keys(value).length > 0}
-                                    <svelte:component this={renderComponent} {...value} isGroupByHeader={true} />
+                                    {@const SvelteComponent = renderComponent}
+                                    <SvelteComponent {...value} isGroupByHeader={true} />
                                 {:else}
-                                    <svelte:component this={renderComponent} {...{value: value}} isGroupByHeader={true} />
+                                    {@const SvelteComponent_1 = renderComponent}
+                                    <SvelteComponent_1 {...{value: value}} isGroupByHeader={true} />
                                 {/if}
                             {:else}
-                                <svelte:component this={renderComponent} {value} isGroupByHeader={true} />
+                                {@const SvelteComponent_2 = renderComponent}
+                                <SvelteComponent_2 {value} isGroupByHeader={true} />
                             {/if}
                             
                             {@const unpaged = groupHeadersUnpaged.find(x => x.groupKey == header.groupKey)}
-                            <svelte:component this={theme.grid.groupby.rowsCount} showing={header.data.length} total={unpaged?.data.length} />
+                            <theme.grid.groupby.rowsCount showing={header.data.length} total={unpaged?.data.length} />
                         {/if}
-                    </svelte:component>
-                </svelte:component>
+                    </theme.grid.groupby.cell>
+                </theme.grid.groupby.container>
                 {#if header.expanded}
                     {#each header.data as row, index (getUniqueKey(row))}
-                        <svelte:component this={theme.grid.body.row} isOdd={(index + 1) % 2 == 1} index={gridData.indexOf(row)} isSelected="{selectedRows.indexOf(row) >= 0}">
+                        <theme.grid.body.row isOdd={(index + 1) % 2 == 1} index={gridData.indexOf(row)} isSelected="{selectedRows.indexOf(row) >= 0}">
                             {#if showCheckboxes}
-                                <svelte:component this={theme.grid.body.checkbox} value={row} index={gridData.indexOf(row)} bind:group={selectedRows} />
+                                <theme.grid.body.checkbox value={row} index={gridData.indexOf(row)} bind:group={selectedRows} />
                             {/if}
                             {#each columns.filter(x => x.visible != false && x.key != groupBy) as col (col.key)}
-                                <svelte:component this={theme.grid.body.cell}>
+                                <theme.grid.body.cell>
                                     {@const value = col.accessor ? col.accessor(row) : row[col.key]}
                                     {@const renderComponent = col.renderComponent ? col.renderComponent : theme.grid.body.content}
                 
                                     {#if typeof value === 'object'}
                                         {#if Object.keys(value).length > 0}
-                                            <svelte:component this={renderComponent} {...value} />
+                                            {@const SvelteComponent_3 = renderComponent}
+                                            <SvelteComponent_3 {...value} />
                                         {:else}
-                                            <svelte:component this={renderComponent} {...{value: value}} />
+                                            {@const SvelteComponent_4 = renderComponent}
+                                            <SvelteComponent_4 {...{value: value}} />
                                         {/if}
                                     {:else}
-                                        <svelte:component this={renderComponent} {value} />
+                                        {@const SvelteComponent_5 = renderComponent}
+                                        <SvelteComponent_5 {value} />
                                     {/if}
-                                </svelte:component>
+                                </theme.grid.body.cell>
                             {/each}
-                        </svelte:component>
+                        </theme.grid.body.row>
                     {/each}
                 {/if}
             {/each}
         {:else}
             {#each gridData as row, index (getUniqueKey(row))}
-                <svelte:component this={theme.grid.body.row} isOdd={(index+1) % 2 == 1} {index} isSelected="{selectedRows.indexOf(row) >= 0}">
+                <theme.grid.body.row isOdd={(index+1) % 2 == 1} {index} isSelected="{selectedRows.indexOf(row) >= 0}">
                     {#if showCheckboxes}
-                        <svelte:component this={theme.grid.body.checkbox} value={row} bind:group={selectedRows} index={index} />
+                        <theme.grid.body.checkbox value={row} bind:group={selectedRows} index={index} />
                     {/if}
                     {#each columns.filter(x => x.visible != false) as col (col.key)}
-                        <svelte:component this={theme.grid.body.cell}>
+                        <theme.grid.body.cell>
                             {@const value = col.accessor ? col.accessor(row) : row[col.key]}
                             {@const renderComponent = col.renderComponent ? col.renderComponent : theme.grid.body.content}
 
                             {#if typeof value === 'object'}
                                 {#if Object.keys(value).length > 0}
-                                    <svelte:component this={renderComponent} {...value} />
+                                    {@const SvelteComponent_6 = renderComponent}
+                                    <SvelteComponent_6 {...value} />
                                 {:else}
-                                    <svelte:component this={renderComponent} {...{value: value}} />
+                                    {@const SvelteComponent_7 = renderComponent}
+                                    <SvelteComponent_7 {...{value: value}} />
                                 {/if}
                             {:else}
-                                <svelte:component this={renderComponent} {value} />
+                                {@const SvelteComponent_8 = renderComponent}
+                                <SvelteComponent_8 {value} />
                             {/if}
-                        </svelte:component>
+                        </theme.grid.body.cell>
                     {/each}
-                </svelte:component>
+                </theme.grid.body.row>
             {/each}
         {/if}
-    </svelte:component>
-</svelte:component>
+    </theme.grid.body.container>
+</theme.grid.container>
