@@ -1,8 +1,12 @@
 <script lang="ts">
     import { PlainTableCssTheme, type GridTheme, PagingData } from '$lib/index.js';
 
-    export let theme:GridTheme = PlainTableCssTheme;
-    export let paging:PagingData = new PagingData();
+    interface Props {
+        theme?: GridTheme;
+        paging?: PagingData;
+    }
+
+    let { theme = PlainTableCssTheme, paging = $bindable(new PagingData()) }: Props = $props();
 </script>
 
-<svelte:component this={theme.paging} bind:paging />
+<theme.paging bind:paging />
