@@ -1,5 +1,9 @@
 <script lang="ts">
-    export let value = "";
+    interface Props {
+        value?: string;
+    }
+
+    let { value = "" }: Props = $props();
 </script>
 
 <span class="block text-sm text-gray-700 dark:text-gray-200">
