@@ -1,13 +1,23 @@
 <script lang="ts">
-    export let title = "";
-    export let onClick = () => {};
-    export let sortable = false;
+    interface Props {
+        title?: string;
+        onClick?: any;
+        sortable?: boolean;
+        children?: import('svelte').Snippet;
+    }
+
+    let {
+        title = "",
+        onClick = () => {},
+        sortable = false,
+        children
+    }: Props = $props();
 </script>
 
-<th class="gc-th {sortable ? 'gc-th--sortable' : ''}" scope="col" on:click={onClick}>
+<th class="gc-th {sortable ? 'gc-th--sortable' : ''}" scope="col" onclick={onClick}>
     <div class="gc-th-col">
         <span class="gc-th-col-title">{title}</span>
-        <slot/>
+        {@render children?.()}
     </div>
 </th>
 
