@@ -1,5 +1,13 @@
+<script lang="ts">
+    interface Props {
+        children?: import('svelte').Snippet;
+    }
+
+    let { children }: Props = $props();
+</script>
+
 <tr class="gc-header-tr">
-    <slot />
+    {@render children?.()}
 </tr>
 
 <style>
