@@ -1,18 +1,28 @@
-<script>
-    export let colspan = 1;
-    export let onToggle = () => {};
-    export let isExpanded = false;
+<script lang="ts">
+    interface Props {
+        colspan?: number;
+        onToggle?: any;
+        isExpanded?: boolean;
+        children?: import('svelte').Snippet;
+    }
+
+    let {
+        colspan = 1,
+        onToggle = () => {},
+        isExpanded = false,
+        children
+    }: Props = $props();
 </script>
 
 <td colspan={colspan}>
     <div>
-        <button on:click={onToggle}>
+        <button onclick={onToggle}>
             {#if isExpanded}
                 <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-chevron-down"><polyline points="6 9 12 15 18 9"></polyline></svg>
             {:else}
                 <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" class="feather feather-chevron-right"><polyline points="9 18 15 12 9 6"></polyline></svg>
             {/if}
         </button>
-        <slot />
+        {@render children?.()}
     </div>
 </td>
