@@ -1,7 +1,12 @@
-<script>
+<script lang="ts">
+    interface Props {
+        children?: import('svelte').Snippet;
+    }
+
+    let { children }: Props = $props();
     // export let isSelected = false;
 </script>
 
 <tr>
-    <slot />
+    {@render children?.()}
 </tr>
