@@ -1,10 +1,20 @@
 <script lang="ts">
-    export let title = "";
-    export let onClick = () => {};
-    export let sortable = false;
+    interface Props {
+        title?: string;
+        onClick?: any;
+        sortable?: boolean;
+        children?: import('svelte').Snippet;
+    }
+
+    let {
+        title = "",
+        onClick = () => {},
+        sortable = false,
+        children
+    }: Props = $props();
 </script>
 
-<th scope="col" on:click={onClick}>
+<th scope="col" onclick={onClick}>
     {title}
-    <slot/>
+    {@render children?.()}
 </th>
