@@ -1,5 +1,13 @@
+<script lang="ts">
+    interface Props {
+        children?: import('svelte').Snippet;
+    }
+
+    let { children }: Props = $props();
+</script>
+
 <div class="overflow-x-scroll">
     <table class="max-w-full min-w-full divide-y divide-gray-200 dark:divide-gray-700">
-        <slot />
+        {@render children?.()}
     </table>
 </div>
