@@ -1,6 +1,10 @@
 <script lang="ts">
-    export let showing = 0;
-    export let total = 0;
+    interface Props {
+        showing?: number;
+        total?: number;
+    }
+
+    let { showing = 0, total = 0 }: Props = $props();
 </script>
 
 <small class="inline-block ml-auto text-gray-400 dark:text-gray-600">{showing} / {total}</small>
