@@ -1,3 +1,11 @@
+<script lang="ts">
+    interface Props {
+        children?: import('svelte').Snippet;
+    }
+
+    let { children }: Props = $props();
+</script>
+
 <tbody class="divide-y divide-gray-200 dark:divide-gray-700">
-    <slot />
+    {@render children?.()}
 </tbody>
