@@ -1,5 +1,13 @@
+<script lang="ts">
+    interface Props {
+        children?: import('svelte').Snippet;
+    }
+
+    let { children }: Props = $props();
+</script>
+
 <td class="h-px whitespace-nowrap">
     <div class="px-6 py-3">
-        <slot />
+        {@render children?.()}
     </div>
 </td>
